Add reply messages

+++ DSL/Resql/centops/send-reply-message.sql
@@ -1,6 +1,6 @@
 WITH inserted_message AS (
   INSERT INTO messages (sender_id, receiver_id, message, type, reply_to_message_id)
-  VALUES (:sender_id, :receiver_id, :message, :type::messages_type, :reply_to_message_id)
+  VALUES (:sender_id, :receiver_id, :message, :type::messages_type, :reply_to_message_id::BIGINT)
   RETURNING id
 )
 INSERT INTO message_status (message_id)
